+++ src/Components/Core/Presentation/Babylon/LearningElements/LearningElementController.ts
@@ -78,6 +78,9 @@ export default class LearningElementController
       this.resetIconScale();
       this.proximityToolTipId = -1;
     }
+    if (isInteractable && this.viewModel.isSpecialFocused) {
+      this.picked();
+    }
   }
 
   private displayTooltip(): number {